consolidate example fibonacci questions; fix editor for rivets bug
