--- exampleCourse/questions/fibonacci/answer.html
@@ -1,3 +1,0 @@
-<p>
-  This question will be graded outside of the PrairieLearn system.
-</p>
