--- exampleCourse/questions/fibonacci/fib_template/Makefile
@@ -1,3 +1,0 @@
-.PHONY: build
-build:
-	tar cvhzf ../fib.tar.gz fib
