--- exampleCourse/questions/fibonacci/fib_template/fib/Makefile
@@ -1,3 +1,0 @@
-.PHONY: upload
-upload:
-	@python .pl_upload.py
