--- exampleCourse/questions/fibonacci/info.json
@@ -1,12 +1,0 @@
-{
-    "uuid": "78ce75d6-780f-4966-bcdf-3ec88180aeac",
-    "title": "Write a Fibonacci function",
-    "topic": "Functions",
-    "tags": ["code"],
-    "clientFiles": ["client.js", "question.html", "answer.html", "fib.py", "fib.tar.gz"],
-    "clientTemplates": ["checkout.py"],
-    "type": "File",
-    "options": {
-        "fileName": "fib.py"
-    }
-}
