--- exampleCourse/questions/fibonacci/question.html
@@ -1,28 +1,0 @@
-<h4>Download your problem</h4>
-<p>
-  <a href="data:text/plain;base64,<%= params.fileData %>" download="<%= params.fileName %>" target="_blank">
-    Download <%= params.fileName %>
-  </a>
-</p>
-<p>
-  Alternatively, <a href="<%= questionFile('checkout.py') %>" target="_blank">download the setup script</a>
-  and run <code>python checkout.py</code>.
-</p>
-<p>
-  Complete the empty <tt>fib</tt> function to recursively compute the <i>n</i>th Fibonacci number.
-</p>
-
-<hr>
-
-<h4>Upload the solution</h4>
-<p>
-  <input type="file" id="fileUpload" name="fileUpload" value="Upload Your Solution">
-</p>
-<p>
-  <a id="downloadFileLink" download="<%= params.fileName %>" style="display:none" target="_blank">
-    Download your solution
-  </a>
-</p>
-<p>
-  If you used the setup script, upload your solution from the command line by running <code>make upload</code>.
-</p>
