+++ exampleCourse/questions/fibonacciEditor/answer.html
@@ -1,3 +1,3 @@
 <p>
-  This question will be graded outside of the PrairieLearn system.
+  No true answer displayed for this question.
 </p>
