+++ exampleCourse/questions/fibonacciEditor/question.html
@@ -1,19 +1,11 @@
-<h4>Problem</h4>
 <p>
-  Write a function <tt>fib</tt> that, given a number <tt>n</tt>, calculates the
-  n<sup>th</sup> Fibonacci number.
+  The Fibonacci numbers are 1, 1, 2, 3, 5, 8, ..., where each
+  number is the sum of the two before it.
+</p>
+<p>
+  Write a Python function <tt>fib</tt> that takes a number <tt>n</tt>
+  and returns the n<sup>th</sup> Fibonacci number.
 </p>
 
-<textarea id="hiddenBuffer" data-instavalue="submittedAnswer.studentCode"></textarea>
-<div id="editor" style="height:500px; width:100%"></div>
-
-<script>
-  var editor = getEditor();
-  editor.getSession().setMode("ace/mode/haskell");
-</script>
-
-<style type="text/css">
-  #hiddenBuffer{
-    display:none;
-  }
-</style>
+<div id="editor" style="height:500px; width:100%; border: 1px solid lightgray;">def fib(n):
+    pass</div>
