--- exampleCourse/questions/fibonacciExternal/answer.html
@@ -1,3 +1,0 @@
-<p>
-  No true answer displayed for this question.
-</p>
