--- exampleCourse/questions/fibonacciExternal/info.json
@@ -1,16 +1,0 @@
-{
-    "uuid": "e04f3972-1657-11e7-93ae-92361f002671",
-    "title": "Fibonacci function in editor, external grading",
-    "topic": "Functions",
-    "tags": ["code"],
-    "clientFiles": ["client.js", "question.html", "answer.html", "submission.html"],
-    "type": "Calculation",
-    "singleVariant": true,
-    "gradingMethod": "External",
-    "externalGradingOptions": {
-          "enabled": true,
-          "autograder": "child_pythondev",
-          "environment": "python-main",
-          "image": "prairielearn/centos7-base"
-    }
-}
