--- exampleCourse/questions/fibonacciExternal/question.html
@@ -1,11 +1,0 @@
-<p>
-  The Fibonacci numbers are 1, 1, 2, 3, 5, 8, ..., where each
-  number is the sum of the two before it.
-</p>
-<p>
-  Write a Python function <tt>fib</tt> that takes a number <tt>n</tt>
-  and returns the n<sup>th</sup> Fibonacci number.
-</p>
-
-<div id="editor" style="height:500px; width:100%; border: 1px solid lightgray;">def fib(n):
-    pass</div>
